moved istio related stuff to an internal package
